Factorize Moon phases periodic terms
This moves periodic terms into their own class. This lets having
`Astronoby::MoonPhases` stay short and clear, instead of being full of
dozens of magic numbers.
