+++ lib/astronoby.rb
@@ -7,6 +7,7 @@ require "astronoby/angles/hms"
 require "astronoby/distance"
 require "astronoby/epoch"
 require "astronoby/astronomical_models/ephemeride_lunaire_parisienne"
+require "astronoby/astronomical_models/moon_phases_periodic_terms"
 require "astronoby/bodies/moon"
 require "astronoby/bodies/sun"
 require "astronoby/coordinates/ecliptic"
